requirements.txt
Flask==3.1.0
tortoise-orm==0.19.0
aiohttp == 3.10.0
uvicorn == 0.15.0
